Change suffix js -> jsx
