+++ zipkin-lens/src/components/Browser/TraceSummary/TraceSummaryBar.jsx
