+++ zipkin-lens/src/components/Browser/TraceSummary/TraceSummaryButtons.jsx
