+++ zipkin-lens/src/components/Common/LoadingOverlay.jsx
