+++ zipkin-lens/src/components/Common/ServiceNameBadge.jsx
