+++ zipkin-lens/src/components/GlobalSearch/ConditionDuration.jsx
